Add Phasors Capability (#15)
* add phasor module foundation

* Convert std::size_t to U64.

* replace ArrayDims

* add radiointerferometryc99

* port delay processor

* add variables to modeb

* add dynamic input to pipeline

* remove phasor manipulation methods

* add phasor module to python interface

* add pi physical constant

* add phasor host based calculation

* add extra dependencies

* add extra dependencies

* add delays output

* loggin quickfix

* logger botch job

* the travesty continues

* fix fringe calculation

* apply fixes

* remove unecessary array

* redo logger

* Add phasor Python interface.

* add phasor module test

* Add Blade version print out.

* Add more logging.

+++ src/modules/channelizer/channelizer.cu
@@ -1,9 +1,10 @@
 #include <cuComplex.h>
+#include <stdint.h>
 
 // 4-point FFT
 
 // TODO: Add multiple formats support.
-template<size_t N, size_t NFFT, size_t NPOLS, size_t NTIME, size_t NCHANS>
+template<uint64_t N, uint64_t NFFT, uint64_t NPOLS, uint64_t NTIME, uint64_t NCHANS>
 __global__ void fft_4pnt(const cuFloatComplex* input, cuFloatComplex* output) {
     const int numThreads = (blockDim.x * gridDim.x) * (NFFT * NPOLS);
     const int threadID = (blockIdx.x * blockDim.x + threadIdx.x) * (NFFT * NPOLS);
